requirements.txt
asgiref==3.6.0
autopep8==2.0.1
bootstrap4==0.1.0
chardet==5.2.0
crispy-bootstrap4==2022.1
distlib==0.3.6
Django==4.2
django-crispy-forms==2.0
django-role-permissions==3.2.0
filelock==3.11.0
pillow==10.2.0
platformdirs==3.2.0
pycodestyle==2.10.0
python-dotenv==1.0.0
reportlab==4.1.0
sqlparse==0.4.3
tzdata==2023.3
virtualenv==20.21.0
